Added logging to `Interop`
- Moved logging logic into `log.py`
- Moved responsibility for configuring logging from `Fuse` to `log()`
- Replaced `print` with `log.info` in `Interop`
- Added some more logging to `Interop`

Fixes #24
