+++ fuse.py
@@ -1,5 +1,5 @@
 import sublime, sublime_plugin, traceback
-import json, threading, time, sys, os, time, subprocess, logging, logging.handlers
+import json, threading, time, sys, os, time, subprocess
 from types import *
 from .interop import *
 from .msg_parser import *
@@ -8,6 +8,7 @@ from .fuse_util import *
 from .go_to_definition import *
 from .build_view import *
 from .version import VERSION
+from .log import log
 
 gFuse = None
 
@@ -232,7 +233,6 @@ class Fuse():
 		self.startFuseEvent.set()
 
 def plugin_loaded():
-	configure_logging()
 	log().info("Loading plugin")
 	log().info("Sublime version '" + sublime.version() + "'")
 	log().info("Fuse plugin version '" + VERSION + "'")
@@ -250,45 +250,6 @@ def plugin_loaded():
 		sublime.active_window().run_command("open_file", {"file":"${packages}/Fuse/UserGuide.txt"})
 		setSetting("fuse_show_user_guide_on_start", False)
 	log().info("Plugin loaded successfully")
-
-def log():
-	return logging.getLogger(__name__)
-
-def userdata_dir():
-	if os.name == "posix":
-		return os.path.join(os.getenv("HOME"), ".fuse")
-	else:
-		return os.path.join(os.getenv("LOCALAPPDATA"), "Fusetools", "Fuse")
-
-def log_dir():
-	return os.path.join(userdata_dir(), "logs")
-
-def log_file():
-	return os.path.join(log_dir(), "fuse.sublimeplugin.log")
-
-def ensure_dir_exists(dir):
-	try:
-		os.makedirs(dir)
-	except FileExistsError:
-		return
-	except Exception as e:
-		sublime.error_message("Could not create directory '" + dir + "'. Please make sure you have the correct permissions: " + str(e))
-
-def configure_logging():
-	log = logging.getLogger(__name__.split(".")[0])
-	if (len(log.handlers) > 0): #Already configured
-		return
-
-	print("logging to " + log_file())
-	ensure_dir_exists(log_dir())
-	handler = logging.handlers.RotatingFileHandler(log_file(), 'a', 500000, 5, "utf-8")
-	formatter = logging.Formatter('%(asctime)s [%(process)d] %(levelname)s %(message)s')
-	handler.setFormatter(formatter)
-	handler.setLevel(logging.INFO)
-	log.setLevel(logging.INFO)
-	log.addHandler(handler)
-	log.info("Finished configuring logging for " + log.name)
-
 
 def fix_osx_path():
 	if str(sublime.platform()) == "osx":
